Moved Guards from cli module to utils module for reuse in contract comparator project
