+++ ds3-autogen-cli/src/main/java/com/spectralogic/autogen/cli/CLI.java
@@ -15,6 +15,7 @@
 
 package com.spectralogic.autogen.cli;
 
+import com.spectralogic.ds3autogen.utils.Guards;
 import org.apache.commons.cli.*;
 
 public class CLI {
